Refactored code for dashboard feature
Codes which were in DashboardController class is moved to
DashboardServiceImpl as the codes are related to logic of the
application.
